move find-related code to a separate crate
