+++ packages/fuel-indexer-plugin/src/lib.rs
@@ -1,3 +1,5 @@
+mod find;
+
 pub mod wasm;
 
 pub mod types {
